Move sponsor-event funcs to events folder, continue events UI
