+++ src/server/nfc/index.test.ts
@@ -1,13 +1,7 @@
 import { ObjectId, MongoError, ObjectID } from 'mongodb';
 import { MongoMemoryServer } from 'mongodb-memory-server';
 import { query } from '../resolvers/helpers';
-import {
-	HackerDbObject,
-	OrganizerDbObject,
-	EventDbObject,
-	CompanyDbObject,
-	TierDbObject,
-} from '../generated/graphql';
+import { HackerDbObject, OrganizerDbObject, EventDbObject } from '../generated/graphql';
 import DB, { Models } from '../models';
 import {
 	checkIfNFCUIDExisted,
@@ -20,8 +14,6 @@ import {
 	shouldWarnRepeatedCheckIn,
 	getEventsAttended,
 	getAttendees,
-	checkIdentityForEvent,
-	assignEventToCompany,
 } from '.';
 
 const event: EventDbObject = {
@@ -66,14 +58,6 @@ const testEvent = ({
 	warnRepeatedCheckins: true,
 } as unknown) as EventDbObject;
 
-const testCompanyId = new ObjectId();
-const testCompany = ({
-	_id: testCompanyId,
-	name: 'test company',
-	tier: { _id: new ObjectId(), name: 'tier', permissions: ['none'] } as TierDbObject,
-	eventsOwned: [],
-} as unknown) as CompanyDbObject;
-
 beforeAll(async () => {
 	try {
 		mongoServer = new MongoMemoryServer();
@@ -85,7 +69,6 @@ beforeAll(async () => {
 		await models.Organizers.insertOne(testOrganizer);
 		await models.Events.insertOne(testEvent);
 		await models.Events.insertOne(event);
-		await models.Companies.insertOne(testCompany);
 	} catch (err) {
 		// eslint-disable-next-line no-console
 		console.error(err);
@@ -584,96 +567,5 @@ describe('Test event model', () => {
 				});
 			});
 		});
-	});
-	describe('checkIdentityForEvent', () => {
-		afterAll(async () => {
-			try {
-				const ret = await models.Events.findOneAndUpdate(
-					{ _id: testEventId },
-					{
-						$set: {
-							owner: null,
-						},
-					}
-				);
-				if (!ret.value) throw new MongoError("Can't set attendees");
-			} catch (err) {
-				// eslint-disable-next-line no-console
-				console.error(err);
-			}
-		});
-		describe('catch error', () => {
-			beforeEach(async () => {
-				try {
-					models.Companies.deleteOne({ _id: testCompanyId });
-				} catch (err) {
-					// eslint-disable-next-line no-console
-					console.error(err);
-				}
-			});
-			afterAll(async () => {
-				try {
-					models.Companies.insertOne(testCompany);
-				} catch (err) {
-					// eslint-disable-next-line no-console
-					console.error(err);
-				}
-			});
-			it('Throws error on non-existent company', async () => {
-				try {
-					await checkIdentityForEvent(
-						testEventId.toHexString(),
-						testCompanyId.toHexString(),
-						models
-					);
-				} catch (err) {
-					expect(err.message).toEqual('Company not found in database');
-				}
-			});
-		});
-
-		it('Return false on non-associated event', async () => {
-			await expect(
-				checkIdentityForEvent(testEventId.toHexString(), testCompanyId.toHexString(), models)
-			).resolves.toEqual(false);
-		});
-		describe('check for true helper', () => {
-			beforeEach(async () => {
-				try {
-					const ret = await models.Events.findOneAndUpdate(
-						{ _id: testEventId },
-						{
-							$set: {
-								owner: testCompany,
-							},
-						}
-					);
-					if (!ret.value) throw new MongoError("Can't set attendees");
-				} catch (err) {
-					// eslint-disable-next-line no-console
-					console.error(err);
-				}
-			});
-			it('Returns true on associated event', async () => {
-				await expect(
-					checkIdentityForEvent(testEventId.toHexString(), testCompanyId.toHexString(), models)
-				).resolves.toEqual(true);
-			});
-		});
-	});
-	describe('assignEventToCompany', () => {
-		it('Sponsor-event association works in the best case', async () => {
-			await assignEventToCompany(testEventId.toHexString(), testCompanyId.toHexString(), models);
-			const companyAfter = await models.Companies.findOne({ _id: testCompanyId });
-			const eventAfter = await models.Events.findOne({ _id: testEventId });
-			if (companyAfter && eventAfter) {
-				expect(companyAfter.eventsOwned).toContain(testEventId.toHexString());
-				if (eventAfter.owner)
-					expect(eventAfter.owner._id.toHexString()).toEqual(testCompanyId.toHexString());
-				else throw new Error('Event not associated with owner');
-			} else throw new Error('Updated company or event not found');
-		});
-
-		// TODO: Additional testing for error throwing
 	});
 });
